pyproject.toml
[project]
name = "ai-whisper-websearch-llm-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "black>=25.1.0",
    "boto3>=1.40.25",
    "faster-whisper>=1.2.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.31",
    "langchain-community>=0.3.29",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.7",
    "langgraph>=0.6.7",
    "lxml>=6.0.1",
    "mcp[cli]>=1.13.1",
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "pyttsx3>=2.99",
    "readability-lxml>=0.8.4.1",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "streamlit>=1.49.1",
    "webrtcvad>=2.0.10",
]
